refactor: split / into routing hub + /landing page
/ is now a pure routing hub that decides where to redirect:
- External homepage configured → redirect there
- Authenticated → /dashboard
- Not authenticated → /landing (built-in landing page)

This fixes several redirect issues:
- No more flash of landing page before external homepage redirect
- No more redirect loop on logout (dashboard → / → dashboard)
- Auth logout callback added to public routes
- Logout sets isLoading=true to prevent shouldRedirect race condition
- Login/register redirect to / (hub) instead of directly to /dashboard
